Added clock algorithm to TLB, added separate tests for TLB, tweaked readonly in vmm
TLB now uses a clock algorithm with a reference bit and a modification bit, miss rate is now improved in most cases. Tests for TLB have been added in a separate file for the time being, a first part checks if the swapping is done as expected, the second part is somewhat random to see a bigger difference in miss rate. In the read and write functions of VMM, the readonly variable is now passed around the different functions that can use it.